Corrects local stereo definition for Sn2 reactions
Sn2 reactions were breaking conversion to local stereo, because the
graph gets reversed while determining the canonical direction. This
means that local stereo is determined with respect to the canonical
direction as well, which makes local stereo depend on canonicalization,
defeating its proper definition.

In the new implementation, parity_evaluator_flip_local_() accepts a flag
indicating that the TS has been reversed. If so, it checks if this is an
Sn2 reaction and applies an extra inversion relative to the canonical
parity.