Rewrites and simplifies graph => geom conversion
Drastically simplifies the geometry generation procedure using the new direct
graph <=> RDKit molecule conversion functionality.

The complexity of not knowing the geometry atom ordering was making things very
complicated and was a constant source of bugs and problems. Hopefully those are
fixed now, but if not they should be much easier to debug.